QTE Plugin
Contains all the information regarding the qte plugin. It can be integrated into any project provided that it is using version 5.0, no later versions are supported. Some additional things that need to be done are all pertaining to the UI in terms of visual effects. The main problem there is that the UI for precision inputs using a material not included so it draws it as a box. This should be updated later